+++ src/main/java/com/teamProject/ezmeal/dao/DeliveryAddressDao.java
@@ -5,7 +5,7 @@ import com.teamProject.ezmeal.domain.DeliveryAddressDto;
 import java.util.List;
 
 public interface DeliveryAddressDao {
-    DeliveryAddressDto defaultAddress(int mbrId) throws Exception;
+    DeliveryAddressDto defaultAddress(Long mbrId) throws Exception;
 
-    List<DeliveryAddressDto> editAddress(int mbrId) throws Exception;
+    List<DeliveryAddressDto> editAddress(Long mbrId) throws Exception;
 }
